feat: add notify manager
